More efficient chunk merging.
We must not read the file contents into memory, especially if we deal
with large amount of data which is the main use case of chunked uploads.

This commit fixes the old behavior and introduces a more efficient merging.
We still read every chunk into memory but this happens in a generator
and we do not merge in-memory so that the memory usage stays constant.